Initial migration of project from Python to SvelteKit.
Reason for the migration is that JS/TS can be used for all logic, instead of having a split between Python and JS. Also, the developer experience is much better, thanks to hot-reloading and possible parallel instances of the web server (at least didn't work out of the box with pywebview). Overall, the developer experience will be more streamlined this way. It's probably easier to export a standalone application/website/app. Modularization of UI components works better, too. So, ultimately many benefits. So far, the migration is mostly complete, besides the data import that will follow in the next step.