Test infrastructure refactoring: reverting sandbox approach.
Turned out the strategy to run the tests in a temporary directory with a project copy didn't work as expected. Instead of spending more time on resolving bugs, increasing the complexity of the setup further, decided to simplify the test setup and try running in parallel to the dev server and present production data.