Big refactoring of current css.
Switched to Sass, split up and refactored styling, so it's dryer and better structured. Now, the glassmorphism parts are defined in a theme as pure design classes, while the html is supposed to have only semantic classes, and a theme integration scss integrates the design classes into semantic classes. This is very clean, reusable, and modular.